fix(grammars): widen uint8_t loop counters in objectscript scanner (CodeQL)
CodeQL cpp/comparison-with-wider-type (4 high-severity alerts, #69-#72) flagged
two loops in the vendored ObjectScript scanner (objectscript_common.h, in both
the _routine and _udl grammar copies) where a uint8_t counter is compared
against an int length: the reverse_marker scan and the html_marker_buffer
reversal. Both lengths are hard-bounded by MARKER_BUFFER_MAX_LEN (30), so the
uint8_t counter can never wrap and the flagged infinite loop is unreachable in
practice — but widening the counters to int removes the pattern and clears the
gate. Provably behavior-neutral (int holds every uint8_t value; length <= 30).

Recorded as a local modification in the vendored grammar MANIFEST so it is
re-applied on the next re-vendor.

+++ internal/cbm/vendored/grammars/objectscript_udl/objectscript_common.h
@@ -332,7 +332,7 @@ static bool ObjectScript_Core_Scanner_lex_marker_fenced_text(
       lexer->mark_end(lexer);
       advance(lexer);
 
-      uint8_t i = 0;
+      int i = 0;
       while (i < reverse_marker_len && !lexer->eof(lexer)
              && lexer->lookahead == reverse_marker[i]) {
         advance(lexer);
@@ -588,7 +588,7 @@ ObjectScript_Core_Scanner_scan(struct ObjectScript_Core_Scanner *scanner,
       return false;
     }
     scanner->js_marker_buffer_reversed_len = scanner->html_marker_buffer_len;
-    for (uint8_t i = 0; i < scanner->html_marker_buffer_len; i++) {
+    for (int i = 0; i < scanner->html_marker_buffer_len; i++) {
       if (scanner->html_marker_buffer[scanner->html_marker_buffer_len - 1 - i] == '[') {
         scanner->js_marker_buffer_reversed[i] = ']';
       }
